api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "neo4j-driver>=5.28.2",
    "requests>=2.32.5",
]
